tree: extract TCP ctrl matching into tree-fabrics.c
Move transport-specific ctrl matching code out of tree.c:
- TCP matching functions and _candidate_init_fabrics() move to
  tree-fabrics.c (compiled only when CONFIG_FABRICS is set)
- _match_ctrl() renamed to _tree_ctrl_match() (non-static) so it
  can be referenced from tree-fabrics.c
- no-fabrics.c gets a _candidate_init_fabrics() stub returning NULL
- streq0()/streqcase0() helpers move to private.h where both tree.c
  and tree-fabrics.c can reach them without an extra include
- struct candidate_args / ctrl_match_t / related declarations move
  to private-fabrics.h; the iface_list field stays conditionally
  compiled on HAVE_NETDB||CONFIG_FABRICS while the rest of the
  struct is unconditional (portability fix for musl/HAVE_NETDB=0)

Tests:
- test/tree.c renamed to test/tree-fabrics.c (fabrics-only); move its
  meson registration under if want_fabrics inside the HAVE_NETDB block
- test/tree-fabrics.c extended with 7 new test functions covering
  PCIe/loop matching, well-known NQN, "none" normalization, RDMA/FC
  config matching, and pagination
- new test/tree.c added with 7 non-fabrics tests for host/subsystem
  creation, deduplication, iteration, and attribute getters; always
  compiled regardless of fabrics config
