+++ libnvme/test/meson.build
@@ -128,6 +128,18 @@ uuid = executable(
 
 test('libnvme - uuid', uuid)
 
+tree = executable(
+    'test-tree',
+    ['tree.c'],
+    dependencies: [
+        config_dep,
+        ccan_dep,
+        libnvme_test_dep,
+    ],
+)
+
+test('libnvme - tree', tree)
+
 if want_fabrics
     uriparser = executable(
         'test-uriparser',
@@ -172,18 +184,20 @@ if conf.get('HAVE_NETDB')
     mock_ifaddrs_env.append('LD_PRELOAD', mock_ifaddrs.full_path())
     mock_ifaddrs_env.set('ASAN_OPTIONS', 'verify_asan_link_order=0')
 
-    tree = executable(
-        'tree',
-        ['tree.c'],
-        dependencies: [
-            config_dep,
-            ccan_dep,
-            libnvme_test_dep,
-        ],
-        link_with: mock_ifaddrs,
-    )
+    if want_fabrics
+        tree_fabrics = executable(
+            'tree-fabrics',
+            ['tree-fabrics.c'],
+            dependencies: [
+                config_dep,
+                ccan_dep,
+                libnvme_test_dep,
+            ],
+            link_with: mock_ifaddrs,
+        )
 
-    test('libnvme - tree', tree, env: mock_ifaddrs_env)
+        test('libnvme - tree-fabrics', tree_fabrics, env: mock_ifaddrs_env)
+    endif
 
     test_util = executable(
         'test-util',
