+++ libnvme/test/tree.c
@@ -1,1210 +1,318 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 /**
  * This file is part of libnvme.
- * Copyright (c) 2023 Daniel Wagner, SUSE LLC
+ * Copyright (c) 2026 Dell Technologies Inc. or its subsidiaries.
+ * Authors: Martin Belanger <Martin.Belanger@dell.com>
+ *
+ * Unit tests for non-fabrics tree operations in libnvme/src/nvme/tree.c:
+ * host/subsystem creation, deduplication, iteration, and attribute getters.
  */
 
 #include <assert.h>
 #include <stdbool.h>
 #include <stdlib.h>
 #include <string.h>
-#include <arpa/inet.h>
-
-#include <ccan/array_size/array_size.h>
+#include <stdio.h>
 
 #include <libnvme.h>
 #include <nvme/private.h>
-#include <nvme/private-fabrics.h>
 
-struct test_data {
-	/* input data */
-	struct libnvmf_context f;
-	const char *subsysname;
+#define HOSTNQN_1 "nqn.2014-08.org.nvmexpress:uuid:aaaaaaaa-1111-1111-1111-aaaaaaaaaaaa"
+#define HOSTID_1  "aaaaaaaa-1111-1111-1111-aaaaaaaaaaaa"
+#define HOSTNQN_2 "nqn.2014-08.org.nvmexpress:uuid:bbbbbbbb-2222-2222-2222-bbbbbbbbbbbb"
+#define HOSTID_2  "bbbbbbbb-2222-2222-2222-bbbbbbbbbbbb"
+#define HOSTNQN_3 "nqn.2014-08.org.nvmexpress:uuid:cccccccc-3333-3333-3333-cccccccccccc"
+#define HOSTID_3  "cccccccc-3333-3333-3333-cccccccccccc"
 
-	/* track controller generated by input data */
-	libnvme_subsystem_t s;
-	libnvme_ctrl_t c;
-	int ctrl_id;
-};
+#define SUBSYSNAME_1 "subsys1"
+#define SUBSYSNQN_1  "nqn.2022-01.com.example:subsys1"
+#define SUBSYSNAME_2 "subsys2"
+#define SUBSYSNQN_2  "nqn.2022-01.com.example:subsys2"
 
-#define DEFAULT_HOSTID "9ba1651a-ed36-11f0-9858-6c1ff71ba506"
-#define DEFAULT_HOSTNQN "nqn.2014-08.org.nvmexpress:uuid:9ba1651a-ed36-11f0-9858-6c1ff71ba506"
-#define DEFAULT_SUBSYSNAME "subsysname"
-#define DEFAULT_SUBSYSNQN "subsysnqn"
-#define SRC_ADDR4 "192.168.56.100"
-#define SRC_ADDR6 "1234:5678:abcd:EF01:1234:5678:abcd:EF01"
-
-#define DEFAULTS(t, a, h, i, s) \
-	.f = { .hostnqn = DEFAULT_HOSTNQN, \
-		.subsysnqn = DEFAULT_SUBSYSNQN,			     \
-		.transport = (t), .traddr = (a), .host_traddr = (h), \
-		.host_iface = (i), .trsvcid = (s), },		     \
-	.subsysname = DEFAULT_SUBSYSNAME
-
-struct test_data test_data[] = {
-	{ DEFAULTS("tcp", "192.168.1.1", "192.168.1.20", NULL, "4420") },
-	{ DEFAULTS("tcp", "192.168.1.1", "192.168.1.20", NULL, "4421") },
-	{ DEFAULTS("tcp", "192.168.1.2", "192.168.1.20", "eth1", "4420") },
-	{ DEFAULTS("tcp", "192.168.1.2", "192.168.1.20", "eth1", "4421") },
-	{ DEFAULTS("rdma", "192.168.1.3", "192.168.1.20", NULL, NULL) },
-	{ DEFAULTS("rdma", "192.168.1.4", "192.168.1.20", NULL, NULL) },
-	{ DEFAULTS("fc",
-		   "nn-0x201700a09890f5bf:pn-0x201900a09890f5bf",
-		   "nn-0x200000109b579ef3:pn-0x100000109b579ef3", NULL, NULL)
-	},
-	{ DEFAULTS("fc",
-		   "nn-0x201700a09890f5bf:pn-0x201900a09890f5bf",
-		   "nn-0x200000109b579ef6:pn-0x100000109b579ef6", NULL, NULL)
-	},
-};
-
-static struct test_data *find_test_data(libnvme_ctrl_t c)
+/**
+ * test_host_dedup - libnvme_lookup_host() must return the same pointer for
+ * the same hostnqn+hostid, and a different pointer for different credentials.
+ */
+static bool test_host_dedup(void)
 {
-	for (int i = 0; i < ARRAY_SIZE(test_data); i++) {
-		struct test_data *d = &test_data[i];
+	struct libnvme_global_ctx *ctx;
+	libnvme_host_t h1, h2, h3;
+	bool pass = true;
 
-		if (d->c == c)
-			return d;
+	printf("test_host_dedup:\n");
+
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
+	assert(ctx);
+
+	h1 = libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
+	assert(h1);
+
+	h2 = libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
+	assert(h2);
+
+	if (h1 != h2) {
+		printf(" - same hostnqn+hostid must return same pointer [FAIL]\n");
+		pass = false;
+	} else {
+		printf(" - same hostnqn+hostid returns same pointer [PASS]\n");
 	}
 
-	return NULL;
-}
+	h3 = libnvme_lookup_host(ctx, HOSTNQN_2, HOSTID_2);
+	assert(h3);
 
-static void show_ctrl(libnvme_ctrl_t c)
-{
-	struct test_data *d = find_test_data(c);
-
-	if (d)
-		printf("ctrl%d ", d->ctrl_id);
-	else
-		printf("      ");
-
-	printf("0x%p: %s %s %s %s %s %s",
-	       c,
-	       libnvme_ctrl_get_subsysnqn(c),
-	       libnvme_ctrl_get_transport(c),
-	       libnvme_ctrl_get_traddr(c),
-	       libnvme_ctrl_get_host_traddr(c),
-	       libnvme_ctrl_get_host_iface(c),
-	       libnvme_ctrl_get_trsvcid(c));
-}
-
-static bool match_ctrl(struct test_data *d, libnvme_ctrl_t c)
-{
-	bool pass = true;
-	const char *trsvid, *host_traddr, *host_iface;
-
-	if (d->c != c)
+	if (h1 == h3) {
+		printf(" - different hostnqn+hostid must return different pointer [FAIL]\n");
 		pass = false;
+	} else {
+		printf(" - different hostnqn+hostid returns different pointer [PASS]\n");
+	}
 
-	if (strcmp(d->f.transport, libnvme_ctrl_get_transport(d->c)))
-		pass = false;
-
-	if (strcmp(d->f.traddr, libnvme_ctrl_get_traddr(d->c)))
-		pass = false;
-
-
-	host_traddr = libnvme_ctrl_get_host_traddr(c);
-	if (d->f.host_traddr &&
-	    (!host_traddr || strcmp(d->f.host_traddr, host_traddr)))
-		pass = false;
-
-	host_iface = libnvme_ctrl_get_host_iface(c);
-	if (d->f.host_iface &&
-	    (!host_iface || strcmp(d->f.host_iface, host_iface)))
-		pass = false;
-
-	trsvid = libnvme_ctrl_get_trsvcid(c);
-	if (d->f.trsvcid &&
-	    (!trsvid || strcmp(d->f.trsvcid, trsvid)))
-		pass = false;
-
-	printf("[%s]", pass? "PASS" : "FAILED");
-
+	libnvme_free_global_ctx(ctx);
 	return pass;
 }
 
-static struct libnvme_global_ctx *create_tree()
+/**
+ * test_hostid_from_hostnqn - When hostid is NULL, libnvme_lookup_host()
+ * must derive the hostid from the UUID embedded in the hostnqn.
+ */
+static bool test_hostid_from_hostnqn(void)
 {
 	struct libnvme_global_ctx *ctx;
 	libnvme_host_t h;
+	const char *hostid;
+	bool pass = true;
 
-	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_DEBUG);
+	printf("test_hostid_from_hostnqn:\n");
+
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
 	assert(ctx);
-	libnvme_get_host(ctx, DEFAULT_HOSTNQN, DEFAULT_HOSTID, &h);
+
+	h = libnvme_lookup_host(ctx, HOSTNQN_1, NULL);
 	assert(h);
 
-	printf("  ctrls created:\n");
-	for (int i = 0; i < ARRAY_SIZE(test_data); i++) {
-		struct test_data *d = &test_data[i];
-
-		assert(!libnvme_get_subsystem(ctx, h, d->subsysname,
-			d->f.subsysnqn, &d->s));
-		assert(d->s);
-		d->c = libnvme_lookup_ctrl(d->s, &d->f, NULL);
-		assert(d->c);
-		d->ctrl_id = i;
-
-		printf("    ");
-		show_ctrl(d->c);
-		match_ctrl(d, d->c);
-		printf("\n");
+	hostid = libnvme_host_get_hostid(h);
+	if (!hostid || strcmp(hostid, HOSTID_1)) {
+		printf(" - hostid derived from hostnqn UUID [FAIL] (got: %s)\n",
+		       hostid ? hostid : "(null)");
+		pass = false;
+	} else {
+		printf(" - hostid derived from hostnqn UUID [PASS]\n");
 	}
-	printf("\n");
 
-	return ctx;
+	libnvme_free_global_ctx(ctx);
+	return pass;
 }
 
-static unsigned int count_entries(struct libnvme_global_ctx *ctx)
+/**
+ * test_host_attrs - hostnqn and hostid getters must return the values used
+ * at creation time.
+ */
+static bool test_host_attrs(void)
 {
+	struct libnvme_global_ctx *ctx;
 	libnvme_host_t h;
-	libnvme_subsystem_t s;
-	libnvme_ctrl_t c;
-	unsigned int i = 0;
+	bool pass = true;
+
+	printf("test_host_attrs:\n");
+
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
+	assert(ctx);
+
+	h = libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
+	assert(h);
+
+	if (!libnvme_host_get_hostnqn(h) ||
+	    strcmp(libnvme_host_get_hostnqn(h), HOSTNQN_1)) {
+		printf(" - hostnqn getter [FAIL]\n");
+		pass = false;
+	} else {
+		printf(" - hostnqn getter [PASS]\n");
+	}
+
+	if (!libnvme_host_get_hostid(h) ||
+	    strcmp(libnvme_host_get_hostid(h), HOSTID_1)) {
+		printf(" - hostid getter [FAIL]\n");
+		pass = false;
+	} else {
+		printf(" - hostid getter [PASS]\n");
+	}
+
+	libnvme_free_global_ctx(ctx);
+	return pass;
+}
+
+/**
+ * test_host_iteration - libnvme_for_each_host() must visit every host
+ * exactly once.
+ */
+static bool test_host_iteration(void)
+{
+	struct libnvme_global_ctx *ctx;
+	libnvme_host_t h;
+	unsigned int count = 0;
+	bool pass = true;
+
+	printf("test_host_iteration:\n");
+
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
+	assert(ctx);
+
+	libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
+	libnvme_lookup_host(ctx, HOSTNQN_2, HOSTID_2);
+	libnvme_lookup_host(ctx, HOSTNQN_3, HOSTID_3);
 
 	libnvme_for_each_host(ctx, h)
-		libnvme_for_each_subsystem(h, s)
-			libnvme_subsystem_for_each_ctrl(s, c)
-				i++;
+		count++;
 
-	return i;
-}
-
-static bool tcp_ctrl_lookup(libnvme_subsystem_t s, struct test_data *d)
-{
-	libnvme_ctrl_t c;
-	bool pass = true;
-	struct libnvmf_context f = d->f;
-
-	f.host_traddr = NULL;
-	f.host_iface = NULL;
-	c = libnvme_lookup_ctrl(s, &f, NULL);
-	printf("%10s %12s %10s -> ", f.trsvcid, "", "");
-	show_ctrl(c);
-	pass &= match_ctrl(d, c);
-	printf("\n");
-
-	if (d->f.host_traddr) {
-		f = d->f;
-		f.host_iface = NULL;
-		c = libnvme_lookup_ctrl(s, &f, NULL);
-		printf("%10s %12s %10s -> ", f.trsvcid, f.host_traddr, "");
-		show_ctrl(c);
-		pass &= match_ctrl(d, c);
-		printf("\n");
+	if (count != 3) {
+		printf(" - expected 3 hosts, got %u [FAIL]\n", count);
+		pass = false;
+	} else {
+		printf(" - 3 hosts found via for_each_host [PASS]\n");
 	}
-
-	if (d->f.host_iface) {
-		f = d->f;
-		f.host_traddr = NULL;
-		c = libnvme_lookup_ctrl(s, &f, NULL);
-		printf("%10s %12s %10s -> ", f.trsvcid, "", f.host_iface);
-		show_ctrl(c);
-		pass &= match_ctrl(d, c);
-		printf("\n");
-	}
-
-	if (d->f.host_iface && d->f.traddr) {
-		f = d->f;
-		c = libnvme_lookup_ctrl(s, &f, NULL);
-		printf("%10s %12s %10s -> ", f.trsvcid,
-		       f.host_traddr, f.host_iface);
-		show_ctrl(c);
-		pass &= match_ctrl(d, c);
-		printf("\n");
-	}
-
-	return pass;
-}
-
-static bool default_ctrl_lookup(libnvme_subsystem_t s, struct test_data *d)
-{
-	libnvme_ctrl_t c;
-	bool pass = true;
-	struct libnvmf_context f = d->f;
-
-	f.host_iface = NULL;
-	f.trsvcid = NULL;
-	c = libnvme_lookup_ctrl(s, &f, NULL);
-	printf("%10s %12s %10s -> ", "", "", "");
-	show_ctrl(c);
-	pass &= match_ctrl(d, c);
-	printf("\n");
-
-	return pass;
-}
-
-static bool ctrl_lookups(struct libnvme_global_ctx *ctx)
-{
-	libnvme_host_t h;
-	libnvme_subsystem_t s;
-	bool pass = true;
-
-	h = libnvme_first_host(ctx);
-	libnvme_get_subsystem(ctx, h, DEFAULT_SUBSYSNAME, DEFAULT_SUBSYSNQN, &s);
-
-	printf("  lookup controller:\n");
-	for (int i = 0; i < ARRAY_SIZE(test_data); i++) {
-		struct test_data *d = &test_data[i];
-
-		printf("%10s %12s %10s    ", "", "", "");
-		show_ctrl(d->c);
-		printf("\n");
-
-		if (!strcmp("tcp", d->f.transport))
-			pass &= tcp_ctrl_lookup(s, d);
-		else
-			pass &= default_ctrl_lookup(s, d);
-
-		printf("\n");
-	}
-
-	return pass;
-}
-
-static bool test_lookup(void)
-{
-	struct libnvme_global_ctx *ctx;
-	bool pass;
-
-	printf("test_lookup:\n");
-
-	ctx = create_tree();
-	pass = count_entries(ctx) == ARRAY_SIZE(test_data);
-	pass &= ctrl_lookups(ctx);
 
 	libnvme_free_global_ctx(ctx);
-
 	return pass;
 }
 
-static bool test_src_addr()
+/**
+ * test_subsystem_dedup - libnvme_lookup_subsystem() must return the same
+ * pointer for the same name+subsysnqn, and a different pointer for different
+ * ones.
+ */
+static bool test_subsystem_dedup(void)
 {
 	struct libnvme_global_ctx *ctx;
-	struct libnvmf_context fctx = {
-		.transport = "tcp",
-		.traddr = "192.168.56.1",
-		.host_traddr = NULL,
-		.host_iface = NULL,
-		.trsvcid = "8009",
-	};
-	bool pass = true;
 	libnvme_host_t h;
-	libnvme_ctrl_t c;
-	libnvme_subsystem_t s;
-	char *src_addr, buffer[100]; /* big enough for IPv6 max length */
+	libnvme_subsystem_t s1, s2, s3;
+	bool pass = true;
 
-	printf("\n"
-	       "test_src_addr:\n");
+	printf("test_subsystem_dedup:\n");
 
-	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_DEBUG);
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
 	assert(ctx);
 
-	libnvme_get_host(ctx, DEFAULT_HOSTNQN, DEFAULT_HOSTID, &h);
+	h = libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
 	assert(h);
 
-	libnvme_get_subsystem(ctx, h, DEFAULT_SUBSYSNAME, DEFAULT_SUBSYSNQN, &s);
-	assert(s);
+	s1 = libnvme_lookup_subsystem(h, SUBSYSNAME_1, SUBSYSNQN_1);
+	assert(s1);
 
-	c = libnvme_lookup_ctrl(s, &fctx, NULL);
-	assert(c);
+	s2 = libnvme_lookup_subsystem(h, SUBSYSNAME_1, SUBSYSNQN_1);
+	assert(s2);
 
-	c->address = NULL;
-	printf(" - Test c->address = NULL                                                       : src_addr = NULL             ");
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (src_addr != NULL) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=NULL should return src_addr=NULL\n");
+	if (s1 != s2) {
+		printf(" - same name+subsysnqn must return same pointer [FAIL]\n");
 		pass = false;
 	} else {
-		printf("[PASS]\n");
+		printf(" - same name+subsysnqn returns same pointer [PASS]\n");
 	}
 
-	c->address = "";
-	printf(" - Test c->address = \"\"                                                         : src_addr = NULL             ");
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (src_addr != NULL) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address="" should return src_addr=NULL\n");
+	s3 = libnvme_lookup_subsystem(h, SUBSYSNAME_2, SUBSYSNQN_2);
+	assert(s3);
+
+	if (s1 == s3) {
+		printf(" - different name+subsysnqn must return different pointer [FAIL]\n");
 		pass = false;
 	} else {
-		printf("[PASS]\n");
+		printf(" - different name+subsysnqn returns different pointer [PASS]\n");
 	}
-
-	c->address = "traddr=192.168.56.1,trsvcid=8009";
-	printf(" - Test c->address = \"%s\"                         : src_addr = NULL             ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (src_addr != NULL) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=NULL\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = "traddr=192.168.56.1,trsvcid=8009,src_addr=" SRC_ADDR4;
-	printf(" - Test c->address = \"%s\" : src_addr = \"" SRC_ADDR4 "\" ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (!src_addr || strcmp(src_addr, SRC_ADDR4)) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=" SRC_ADDR4 "\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = "traddr=192.168.56.1,src_addr=" SRC_ADDR4 ",trsvcid=8009";
-	printf(" - Test c->address = \"%s\" : src_addr = \"" SRC_ADDR4 "\" ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (!src_addr || strcmp(src_addr, SRC_ADDR4)) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=" SRC_ADDR4 "\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = "traddr=1234::abcd,trsvcid=8009,src_addr=" SRC_ADDR6;
-	printf(" - Test c->address = \"%s\"       : src_addr = \"" SRC_ADDR6 "\" ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (!src_addr || strcmp(src_addr, SRC_ADDR6)) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=" SRC_ADDR6 "\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = "traddr=1234::abcd,src_addr=" SRC_ADDR6 ",trsvcid=8009";
-	printf(" - Test c->address = \"%s\"       : src_addr = \"" SRC_ADDR6 "\" ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (!src_addr || strcmp(src_addr, SRC_ADDR6)) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=" SRC_ADDR6 "\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = "traddr=1234::abcd,trsvcid=8009,src_addr=" SRC_ADDR6 "%scope";
-	printf(" - Test c->address = \"%s\" : src_addr = \"" SRC_ADDR6 "\" ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (!src_addr || strcmp(src_addr, SRC_ADDR6)) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=" SRC_ADDR6 "\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = "traddr=1234::abcd,src_addr=" SRC_ADDR6 "%scope,trsvcid=8009";
-	printf(" - Test c->address = \"%s\" : src_addr = \"" SRC_ADDR6 "\" ", c->address);
-	src_addr = libnvme_ctrl_get_src_addr(c, buffer, sizeof(buffer));
-	if (!src_addr || strcmp(src_addr, SRC_ADDR6)) {
-		printf("[FAIL]\n");
-		fprintf(stderr,
-			"libnvme_ctrl_get_src_addr() c->address=%s should return src_addr=" SRC_ADDR6 "\n",
-			c->address);
-		pass = false;
-	} else {
-		printf("[PASS]\n");
-	}
-
-	c->address = NULL; /* Needed to avoid freeing non-malloced memory (see above) */
 
 	libnvme_free_global_ctx(ctx);
-
 	return pass;
 }
 
-struct ctrl_args {
-	struct libnvmf_context f;
-	const char *address;
-};
-
-static void set_ctrl_args(struct ctrl_args *args,
-			  const char *transport,
-			  const char *traddr,
-			  const char *trsvcid,
-			  const char *host_traddr,
-			  const char *host_iface,
-			  const char *address,
-			  const char *subsysnqn)
-{
-	args->f.transport   = transport;
-	args->f.traddr      = traddr;
-	args->f.trsvcid     = trsvcid;
-	args->f.host_traddr = host_traddr;
-	args->f.host_iface  = host_iface;
-	args->address     = address;
-	args->f.subsysnqn   = subsysnqn;
-}
-
-static bool ctrl_match(const char *tag,
-		       int reference_id,
-		       int candidate_id,
-		       struct ctrl_args *reference,
-		       struct ctrl_args *candidate,
-		       bool should_match)
+/**
+ * test_subsystem_attrs - subsysnqn and name getters must return the values
+ * used at creation time.
+ */
+static bool test_subsystem_attrs(void)
 {
 	struct libnvme_global_ctx *ctx;
 	libnvme_host_t h;
-	libnvme_ctrl_t reference_ctrl; /* Existing controller (from sysfs) */
-	libnvme_ctrl_t candidate_ctrl;
-	libnvme_ctrl_t found_ctrl;
 	libnvme_subsystem_t s;
+	bool pass = true;
 
-	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_INFO);
+	printf("test_subsystem_attrs:\n");
+
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
 	assert(ctx);
 
-	libnvme_get_host(ctx, DEFAULT_HOSTNQN, DEFAULT_HOSTID, &h);
+	h = libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
 	assert(h);
 
-	assert(!libnvme_get_subsystem(ctx, h, DEFAULT_SUBSYSNAME,
-		 reference->f.subsysnqn ?
-			reference->f.subsysnqn : DEFAULT_SUBSYSNQN,
-		&s));
+	s = libnvme_lookup_subsystem(h, SUBSYSNAME_1, SUBSYSNQN_1);
 	assert(s);
 
-	reference_ctrl = libnvme_lookup_ctrl(s, &reference->f, NULL);
-	assert(reference_ctrl);
-	reference_ctrl->name = "nvme1";  /* fake the device name */
-	if (reference->address) {
-		reference_ctrl->address = (char *)reference->address;
-	}
-
-	/* libnvme_ctrl_find() MUST BE RUN BEFORE libnvme_lookup_ctrl() */
-	found_ctrl = libnvme_ctrl_find(s, &candidate->f);
-
-	candidate_ctrl = libnvme_lookup_ctrl(s, &candidate->f, NULL);
-
-	if (should_match) {
-		if (candidate_ctrl != reference_ctrl) {
-			printf("%s-%d-%d: Candidate (%s, %s, %s, %s, %s, %s) failed to match (%s, %s, %s, %s, %s, %s, %s)\n",
-			       tag, reference_id, candidate_id,
-			       candidate->f.transport, candidate->f.traddr,
-			       candidate->f.trsvcid, candidate->f.subsysnqn,
-			       candidate->f.host_traddr,
-			       candidate->f.host_iface,
-			       reference->f.transport, reference->f.traddr,
-			       reference->f.trsvcid, reference->f.subsysnqn,
-			       reference->f.host_traddr,
-			       reference->f.host_iface, reference->address);
-			return false;
-		}
-
-		if (!found_ctrl) {
-			printf("%s-%d-%d: Candidate (%s, %s, %s, %s, %s, %s) failed to find controller\n",
-			       tag, reference_id, candidate_id,
-			       candidate->f.transport, candidate->f.traddr,
-			       candidate->f.trsvcid, candidate->f.subsysnqn,
-			       candidate->f.host_traddr,
-			       candidate->f.host_iface);
-			return false;
-		}
+	if (!libnvme_subsystem_get_name(s) ||
+	    strcmp(libnvme_subsystem_get_name(s), SUBSYSNAME_1)) {
+		printf(" - subsystem name getter [FAIL]\n");
+		pass = false;
 	} else {
-		if (candidate_ctrl == reference_ctrl) {
-			printf("%s-%d-%d: Candidate (%s, %s, %s, %s, %s, %s) should not match (%s, %s, %s, %s, %s, %s, %s)\n",
-			       tag, reference_id, candidate_id,
-			       candidate->f.transport, candidate->f.traddr,
-			       candidate->f.trsvcid, candidate->f.subsysnqn,
-			       candidate->f.host_traddr,
-			       candidate->f.host_iface,
-			       reference->f.transport, reference->f.traddr,
-			       reference->f.trsvcid, reference->f.subsysnqn,
-			       reference->f.host_traddr,
-			       reference->f.host_iface, reference->address);
-			return false;
-		}
-
-		if (found_ctrl) {
-			printf("%s-%d-%d: Candidate (%s, %s, %s, %s, %s, %s) should not have found controller. found_ctrl=%p reference=%p\n",
-			       tag, reference_id, candidate_id,
-			       candidate->f.transport, candidate->f.traddr,
-			       candidate->f.trsvcid, candidate->f.subsysnqn,
-			       candidate->f.host_traddr,
-			       candidate->f.host_iface,
-			       found_ctrl, reference_ctrl);
-			return false;
-		}
+		printf(" - subsystem name getter [PASS]\n");
 	}
 
-	/* Set the faked data back to NULL before freeing the tree */
-	reference_ctrl->name = NULL;
-	reference_ctrl->address = NULL;
+	if (!libnvme_subsystem_get_subsysnqn(s) ||
+	    strcmp(libnvme_subsystem_get_subsysnqn(s), SUBSYSNQN_1)) {
+		printf(" - subsysnqn getter [FAIL]\n");
+		pass = false;
+	} else {
+		printf(" - subsysnqn getter [PASS]\n");
+	}
 
 	libnvme_free_global_ctx(ctx);
-
-	return true;
-}
-
-/**
- * test_ctrl_match_fc - Test that we can look up FC controllers
- *
- * @return true when all tests have passed. false otherwise.
- */
-static bool test_ctrl_match_fc(void)
-{
-	bool pass = true;
-	struct ctrl_args reference = {0};
-	struct ctrl_args candidate = {0};
-
-	printf("test_ctrl_match_fc:\n");
-
-	/*******************************************************************/
-	/* Reference ID 1 */
-	set_ctrl_args(&reference, "fc", "21:00:00:e0:8b:05:05:01", "4420", "21:00:00:e0:8b:05:05:20", NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", "21:00:00:e0:8b:05:05:20", NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 1, 0, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 1, 1, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("FC", 1, 2, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "192.168.2.2", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("FC", 1, 3, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 1, 4, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", "21:00:00:e0:8b:05:05:21", NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 1, 5, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "8009", NULL, "", NULL, NULL);
-	pass &= ctrl_match("FC", 1, 6, &reference, &candidate, false);
-
-
-	/*******************************************************************/
-	/* Reference ID 2 */
-	set_ctrl_args(&reference, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", "21:00:00:e0:8b:05:05:20", NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 2, 0, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 2, 1, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("FC", 2, 2, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "192.168.2.2", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("FC", 2, 3, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 2, 4, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "8009", NULL, "", NULL, NULL);
-	pass &= ctrl_match("FC", 2, 5, &reference, &candidate, false);
-
-
-	/*******************************************************************/
-	/* Reference ID 3 */
-	set_ctrl_args(&reference, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, "eth0", NULL, NULL);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", "21:00:00:e0:8b:05:05:20", NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 3, 0, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 3, 1, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("FC", 3, 2, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "fc", "192.168.2.2", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("FC", 3, 3, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("FC", 3, 4, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "fc", "21:00:00:e0:8b:05:05:01", "8009", NULL, "", NULL, NULL);
-	pass &= ctrl_match("FC", 3, 5, &reference, &candidate, false);
-
 	return pass;
 }
 
 /**
- * test_ctrl_match_rdma - Test that we can look up RDMA controllers
- *
- * @return true when all tests have passed. false otherwise.
+ * test_subsystem_iteration - libnvme_for_each_subsystem() must visit every
+ * subsystem exactly once.
  */
-static bool test_ctrl_match_rdma(void)
-{
-	bool pass = true;
-	struct ctrl_args reference = {0};
-	struct ctrl_args candidate = {0};
-
-	printf("test_ctrl_match_rdma:\n");
-
-	/*******************************************************************/
-	/* Reference ID 1 */
-	set_ctrl_args(&reference, "rdma", "192.168.2.1", "4420", "192.168.1.20", NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 0, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 1, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 2, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.2", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 3, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 4, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 5, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "8009", NULL, "", NULL, NULL);
-	pass &= ctrl_match("RDMA", 1, 6, &reference, &candidate, false);
-
-
-	/*******************************************************************/
-	/* Reference ID 2 */
-	set_ctrl_args(&reference, "rdma", "192.168.2.1", "4420", NULL, NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 2, 0, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 2, 1, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("RDMA", 2, 2, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.2", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("RDMA", 2, 3, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 2, 4, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "8009", NULL, "", NULL, NULL);
-	pass &= ctrl_match("RDMA", 2, 5, &reference, &candidate, false);
-
-
-	/*******************************************************************/
-	/* Reference ID 3 */
-	set_ctrl_args(&reference, "rdma", "192.168.2.1", "4420", NULL, "eth0", NULL, NULL);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 3, 0, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 3, 1, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("RDMA", 3, 2, &reference, &candidate, true);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.2", "4420", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("RDMA", 3, 3, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("RDMA", 3, 4, &reference, &candidate, false);
-
-	set_ctrl_args(&candidate, "rdma", "192.168.2.1", "8009", NULL, "", NULL, NULL);
-	pass &= ctrl_match("RDMA", 3, 5, &reference, &candidate, false);
-
-	return pass;
-}
-
-/**
- * test_ctrl_match_tcp - Test that we can look up TCP controllers
- *
- * @note: The mocked getifaddrs() returns 2 interface entries with the
- *        following addresses. Therefore the tests must use IP addresses
- *        that match these.
- *
- *        eth0
- *         \_ 192.168.1.20
- *         \_ fe80::dead:beef
- *
- *        lo
- *         \_ 127.0.0.1
- *         \_ ::1
- *
- * @return true when all tests have passed. false otherwise.
- */
-static bool test_ctrl_match_tcp()
-{
-	bool pass = true;
-	struct ctrl_args reference = {0};
-	struct ctrl_args candidate = {0};
-
-	printf("\n"
-	       "test_ctrl_match_tcp:\n");
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 1 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 4, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 5, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 6, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 7, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 1, 8, &reference, &candidate, true);
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 2 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 2, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 3 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 3, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 4 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 4, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 5 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 1, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 2, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 3, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 4, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 5, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 6 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", NULL, NULL, "traddr=123.123.123.123,trsvcid=8009,src_addr=192.168.1.20", NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 6, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv4: Reference ID 7 */
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", NULL, NULL, "traddr=123.123.123.123,trsvcid=8009,src_addr=127.0.0.1", NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 1, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 2, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 3, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 6, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv4", 7, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 1 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 4, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 5, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 6, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 7, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 1, 8, &reference, &candidate, true);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 2 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 2, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 3 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 3, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 4 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 4, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 5 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 1, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 2, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 3, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 4, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 5, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 6 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", NULL, NULL, "traddr=aaaa::bbbb,trsvcid=8009,src_addr=fe80::dead:beef", NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 6, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 6, 8, &reference, &candidate, false);
-
-	/*******************************************************************/
-	/* IPv6: Reference ID 7 */
-	set_ctrl_args(&reference, "tcp", "aaaa::bbbb", "8009", NULL, NULL, "traddr=aaaa::bbbb,trsvcid=8009,src_addr=::1", NULL);
-
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 1, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 2, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 3, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", NULL, NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 4, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "eth0", NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 5, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 6, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:beef", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 7, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "aaaa::bbbb", "8009", "fe80::dead:cafe", "lo", NULL, NULL);
-	pass &= ctrl_match("IPv6", 7, 8, &reference, &candidate, false);
-
-	return pass;
-}
-
-static bool ctrl_config_match(const char *tag,
-			      int reference_id,
-			      int candidate_id,
-			      struct ctrl_args *reference,
-			      struct ctrl_args *candidate,
-			      bool should_match)
+static bool test_subsystem_iteration(void)
 {
 	struct libnvme_global_ctx *ctx;
-	bool match;
 	libnvme_host_t h;
-	libnvme_ctrl_t reference_ctrl; /* Existing controller (from sysfs) */
 	libnvme_subsystem_t s;
+	unsigned int count = 0;
+	bool pass = true;
 
-	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_INFO);
+	printf("test_subsystem_iteration:\n");
+
+	ctx = libnvme_create_global_ctx(stdout, LIBNVME_LOG_ERR);
 	assert(ctx);
 
-	libnvme_get_host(ctx, DEFAULT_HOSTNQN, DEFAULT_HOSTID, &h);
+	h = libnvme_lookup_host(ctx, HOSTNQN_1, HOSTID_1);
 	assert(h);
 
-	assert(!libnvme_get_subsystem(ctx, h, DEFAULT_SUBSYSNAME,
-		reference->f.subsysnqn ?
-			reference->f.subsysnqn : DEFAULT_SUBSYSNQN,
-		&s));
-	assert(s);
+	libnvme_lookup_subsystem(h, SUBSYSNAME_1, SUBSYSNQN_1);
+	libnvme_lookup_subsystem(h, SUBSYSNAME_2, SUBSYSNQN_2);
 
-	reference_ctrl = libnvme_lookup_ctrl(s, &reference->f, NULL);
-	assert(reference_ctrl);
-	reference_ctrl->name = "nvme1";  /* fake the device name */
-	if (reference->address) {
-		reference_ctrl->address = (char *)reference->address;
-	}
+	libnvme_for_each_subsystem(h, s)
+		count++;
 
-	match = _libnvme_ctrl_match_config(reference_ctrl, &candidate->f);
-
-	if (should_match) {
-		if (!match) {
-			printf("%s-%d-%d: Failed to match config for Candidate (%s, %s, %s, %s, %s, %s)\n",
-			       tag, reference_id, candidate_id,
-			       candidate->f.transport, candidate->f.traddr,
-			       candidate->f.trsvcid, candidate->f.subsysnqn,
-			       candidate->f.host_traddr,
-			       candidate->f.host_iface);
-			return false;
-		}
+	if (count != 2) {
+		printf(" - expected 2 subsystems, got %u [FAIL]\n", count);
+		pass = false;
 	} else {
-		if (match) {
-			printf("%s-%d-%d: Config should not have matched for Candidate (%s, %s, %s, %s, %s, %s)\n",
-			       tag, reference_id, candidate_id,
-			       candidate->f.transport, candidate->f.traddr,
-			       candidate->f.trsvcid, candidate->f.subsysnqn,
-			       candidate->f.host_traddr,
-			       candidate->f.host_iface);
-			return false;
-		}
+		printf(" - 2 subsystems found via for_each_subsystem [PASS]\n");
 	}
-
-	/* Set the faked data back to NULL before freeing the tree */
-	reference_ctrl->name = NULL;
-	reference_ctrl->address = NULL;
 
 	libnvme_free_global_ctx(ctx);
-
-	return true;
-}
-
-static bool test_ctrl_config_match()
-{
-	bool pass = true;
-	struct ctrl_args reference = {0};
-	struct ctrl_args candidate = {0};
-
-	printf("\n"
-	       "test_ctrl_config_match:\n");
-
-	set_ctrl_args(&reference, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, NULL, NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 0, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", NULL, NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 1, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "eth0", NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 2, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "eth0", NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 3, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", NULL, NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 4, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "eth0", NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 5, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", NULL, "lo", NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 6, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.20", "lo", NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 7, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, NULL);
-	pass &= ctrl_config_match("IPv4", 1, 8, &reference, &candidate, true);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, "hello");
-	pass &= ctrl_config_match("IPv4", 1, 9, &reference, &candidate, false);
-	set_ctrl_args(&candidate, "tcp", "123.123.123.123", "8009", "192.168.1.21", "lo", NULL, DEFAULT_SUBSYSNQN);
-	pass &= ctrl_config_match("IPv4", 1, 9, &reference, &candidate, true);
-
 	return pass;
 }
 
-
-/**
- * This test module uses a mocked ifaddrs library (mock-ifaddrs.c)
- * such that there are 2 fake interfaces (eth0 and lo) with the
- * following IP addresses:
- *
- *    - eth0
- *      \_ IPv4: 192.168.1.20
- *      \_ IPv6: fe80::dead:beef
- *
- *    - lo
- *      \_ IPv4: 127.0.0.1
- *      \_ IPv6: ::1
- */
 int main(int argc, char *argv[])
 {
 	bool pass = true;
 
-	pass &= test_lookup();
-	pass &= test_src_addr();
-	pass &= test_ctrl_match_fc();
-	pass &= test_ctrl_match_rdma();
-	pass &= test_ctrl_match_tcp();
-	pass &= test_ctrl_config_match();
+	pass &= test_host_dedup();
+	pass &= test_hostid_from_hostnqn();
+	pass &= test_host_attrs();
+	pass &= test_host_iteration();
+	pass &= test_subsystem_dedup();
+	pass &= test_subsystem_attrs();
+	pass &= test_subsystem_iteration();
 
 	fflush(stdout);
-
 	exit(pass ? EXIT_SUCCESS : EXIT_FAILURE);
 }
